+++ src/main/java/com/example/routefinder/GraphLink.java
@@ -2,10 +2,10 @@ package com.example.routefinder;
 
 public class GraphLink {
     public NodeWithCost<?> destNode; //Could also store source node if required
-    public int cost;
+    public double cost;
 
     //Other link attributes could be similarly stored
-    public GraphLink(NodeWithCost<?> destNode,int cost) {
+    public GraphLink(NodeWithCost<?> destNode,double cost) {
         this.destNode=destNode;
         this.cost=cost;
     }
