+++ src/main/java/com/example/routefinder/NodeWithCost.java
@@ -6,7 +6,7 @@ public class NodeWithCost<T> {
 
     public T data;
     public List<GraphLink> adjList=new ArrayList<>(); //Could use any List implementation
-    public int nodeValue = Integer.MAX_VALUE;
+    public double nodeValue = Double.MAX_VALUE;
 
     public NodeWithCost(T data) {
         this.data=data;
@@ -19,11 +19,11 @@ public class NodeWithCost<T> {
                 '}';
     }
 
-    public void connectToNodeDirected(NodeWithCost<T> destNode,int cost) {
+    public void connectToNodeDirected(NodeWithCost<T> destNode,double cost) {
         adjList.add(new GraphLink(destNode,cost));
     }
 
-    public void connectToNodeUndirected(NodeWithCost<T> destNode,int cost) {
+    public void connectToNodeUndirected(NodeWithCost<T> destNode,double cost) {
         adjList.add(new GraphLink(destNode,cost));
         destNode.adjList.add(new GraphLink(this,cost));
     }
